api-server: move metrics and add metrics for data service api
