+++ packages/api-server/src/plugins/services/tidb-data-service/index.ts
@@ -1,5 +1,8 @@
 import DigestClient from "digest-fetch";
+import {FastifyBaseLogger} from "fastify";
 import fp from "fastify-plugin";
+import {DateTime} from "luxon";
+import {countAPIRequest, dataServiceRequestCounter, dataServiceRequestTimer, measure} from "../../../metrics";
 
 declare module 'fastify' {
   interface FastifyInstance {
@@ -10,6 +13,7 @@ declare module 'fastify' {
 export default fp(async (app) => {
   if (app.config.TIDB_CLOUD_DATA_SERVICE_APP_ID) {
     app.decorate('tidbDataService', new TiDBDataService(
+      app.log,
       app.config.TIDB_CLOUD_DATA_SERVICE_APP_ID,
       app.config.TIDB_CLOUD_DATA_SERVICE_PUBLIC_KEY,
       app.config.TIDB_CLOUD_DATA_SERVICE_PRIVATE_KEY,
@@ -25,6 +29,7 @@ export class TiDBDataService {
   private readonly baseURL: string;
 
   constructor(
+    readonly logger: FastifyBaseLogger,
     appId: string,
     publicKey: string,
     privateKey: string,
@@ -34,9 +39,21 @@ export class TiDBDataService {
   }
 
   async query(endpointPath: string, params: Record<string, any>) {
-    return await this.client.fetch(`${this.baseURL}/${endpointPath}`, {
-      params
+    const counter = dataServiceRequestCounter;
+    const timer = dataServiceRequestTimer.labels({ api: endpointPath });
+
+    return await countAPIRequest(counter, endpointPath, async () => {
+      return await measure(timer, async () => {
+        const startTime = DateTime.now();
+        const res = await this.client.fetch(`${this.baseURL}/${endpointPath}`, {
+          params
+        });
+
+        const endTime = DateTime.now();
+        const duration = endTime.diff(startTime, 'seconds').seconds;
+        this.logger.info(`✅ Finished querying ${endpointPath} on TiDB Data Service, cost: ${duration} s.`);
+        return res;
+      });
     });
   }
-
 }
